[NFC] Extract common code for keying off of APInts
